requirements.txt
torch
transformers
accelerate
qwen-vl-utils
sentence-transformers
faiss-cpu
numpy
streamlit
pymupdf
